requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
streamlit==1.32.2
python-multipart==0.0.9
Pillow==10.2.0
transformers==4.38.2
torch==2.2.1
spacy==3.7.4
python-dateutil==2.8.2
pytesseract==0.3.10
openai==1.12.0 